requirements.txt
fastapi
uvicorn
bcrypt
mysql-connector-python
pydantic
python-dotenv
